+++ activemodel/lib/active_model/error.rb
@@ -7,7 +7,7 @@ module ActiveModel
   #
   # Represents one single error
   class Error
-    CALLBACKS_OPTIONS = [:if, :unless, :on, :allow_nil, :allow_blank, :strict]
+    CALLBACKS_OPTIONS = [:if, :unless, :on, :allow_nil, :allow_blank, :strict, :skip_on]
     MESSAGE_OPTIONS = [:message]
 
     class_attribute :i18n_customize_full_message, default: false