+++ activemodel/CHANGELOG.md
@@ -19,4 +19,21 @@
 
     *Jonathan Hefner*
 
+*   Add `skip_on` option to Active Model Validation Options.
+    this option allow to skip specific validation in specific context.
+
+      ```ruby
+      class MyModel
+        include ActiveModel::Validations
+
+        validates :name, presence: true, skip_on: :custom_context
+      end
+
+      my_model = MyModel.new
+      my_model.valid? # => false
+      my_model.valid?(:custom_context) # => true
+      ```
+
+    *Ahmed Bin Shamlh*
+
 Please check [7-1-stable](https://github.com/rails/rails/blob/7-1-stable/activemodel/CHANGELOG.md) for previous changes.